requirements.txt
Django
djangorestframework
djangorestframework-jwt
django-cors-headers
psycopg2